% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.add_datatype_specific_info}
\alias{.add_datatype_specific_info}
\title{Add Datatype-Specific Information to Summary}
\usage{
.add_datatype_specific_info(info_data, datatype, options, context)
}
\description{
Add Datatype-Specific Information to Summary
}
\keyword{internal}
