% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{.convert_param_by_datatype}
\alias{.convert_param_by_datatype}
\title{Convert Parameter by Data Type}
\usage{
.convert_param_by_datatype(value, datatype)
}
\arguments{
\item{value}{The parameter value to convert.}

\item{datatype}{Character. The target data type.}
}
\value{
Converted value.
}
\description{
Converts a parameter value based on its specified data type.
}
\keyword{internal}
