% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.create_biological_context_if_needed}
\alias{.create_biological_context_if_needed}
\title{Create Biological Context If Needed}
\usage{
.create_biological_context_if_needed(datatype, seed)
}
\description{
Create Biological Context If Needed
}
\keyword{internal}
