% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.create_internal_context}
\alias{.create_internal_context}
\title{Create Internal Genome Context}
\usage{
.create_internal_context(seed)
}
\description{
Create Internal Genome Context
}
\keyword{internal}
