% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.create_output_directories}
\alias{.create_output_directories}
\title{Create Output Directories}
\usage{
.create_output_directories(cmd_config, user_params, show_messages)
}
\description{
Create Output Directories
}
\keyword{internal}
