% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.create_temp_file_path}
\alias{.create_temp_file_path}
\title{Create Temporary File Path with Smart Naming}
\usage{
.create_temp_file_path(datatype, options, compress)
}
\description{
Create Temporary File Path with Smart Naming
}
\keyword{internal}
