% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{.create_tool_from_config}
\alias{.create_tool_from_config}
\title{Create Tool Object from Configuration}
\usage{
.create_tool_from_config(config, version, base_dir)
}
\description{
Create Tool Object from Configuration
}
\keyword{internal}
