% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.determine_compression}
\alias{.determine_compression}
\title{Determine Final Compression Setting}
\usage{
.determine_compression(output_file, datatype, compress)
}
\description{
Determine Final Compression Setting
}
\keyword{internal}
