% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.determine_log_directory}
\alias{.determine_log_directory}
\title{Determine Log Directory}
\usage{
.determine_log_directory(log_dir, work_dir, user_params, output_defs)
}
\description{
Determine Log Directory
}
\keyword{internal}
