% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{.diagnose_tool_installation}
\alias{.diagnose_tool_installation}
\title{Diagnose Tool Installation Issues (Detailed)}
\usage{
.diagnose_tool_installation(tool, version, base_dir, show_details = TRUE)
}
\description{
Diagnose Tool Installation Issues (Detailed)
}
\keyword{internal}
