% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.ensure_output_directory}
\alias{.ensure_output_directory}
\title{Ensure Output Directory Exists}
\usage{
.ensure_output_directory(output_file)
}
\description{
Ensure Output Directory Exists
}
\keyword{internal}
