% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.execute_shell_command}
\alias{.execute_shell_command}
\title{Execute Shell Command}
\usage{
.execute_shell_command(command, env_path, work_dir, log_file, show_output)
}
\description{
Execute Shell Command
}
\keyword{internal}
