% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{.extract_version_from_config}
\alias{.extract_version_from_config}
\title{Extract Version from Tool Configuration}
\usage{
.extract_version_from_config(config)
}
\description{
Extract Version from Tool Configuration
}
\keyword{internal}
