% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.find_tool_env}
\alias{.find_tool_env}
\title{Find Tool Environment Path}
\usage{
.find_tool_env(tool_name, tool_config)
}
\arguments{
\item{tool_name}{Character. Name of the tool.}

\item{tool_config}{List. Tool configuration.}
}
\value{
Character. Path to tool environment or NULL if not found.
}
\description{
Find the installation path for a tool environment.
}
\keyword{internal}
