% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_memory}
\alias{.format_memory}
\title{Format Memory Size}
\usage{
.format_memory(bytes)
}
\arguments{
\item{bytes}{Numeric. Memory size in bytes.}
}
\value{
Character. Formatted memory size.
}
\description{
Formats memory size in bytes to human-readable format.
}
\keyword{internal}
