% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_bed}
\alias{.generate_bed}
\title{Generate Mock BED File}
\usage{
.generate_bed(output_file, n_records, options, genome_context = NULL, compress)
}
\description{
Generate Mock BED File
}
\keyword{internal}
