% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_example_filename}
\alias{.generate_example_filename}
\title{Generate Example Filename Based on Context}
\usage{
.generate_example_filename(
  datatype,
  param_name,
  input_output,
  compress = FALSE
)
}
\description{
Generate Example Filename Based on Context
}
\keyword{internal}
