% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_fasta}
\alias{.generate_fasta}
\title{Generate Mock FASTA File}
\usage{
.generate_fasta(
  output_file,
  n_records,
  options,
  genome_context = NULL,
  compress
)
}
\description{
Generate Mock FASTA File
}
\keyword{internal}
