% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_fastq}
\alias{.generate_fastq}
\title{Generate Mock FASTQ File}
\usage{
.generate_fastq(
  output_file,
  n_records,
  options,
  genome_context = NULL,
  compress
)
}
\description{
Generate Mock FASTQ File
}
\keyword{internal}
