% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_gene_positions}
\alias{.generate_gene_positions}
\title{Generate Gene Positions for Genome Context}
\usage{
.generate_gene_positions(chromosomes, chr_lengths, gene_density, seed)
}
\description{
Generate Gene Positions for Genome Context
}
\keyword{internal}
