% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_gff}
\alias{.generate_gff}
\title{Generate Mock GFF File}
\usage{
.generate_gff(output_file, n_records, options, genome_context = NULL, compress)
}
\description{
Generate Mock GFF File
}
\keyword{internal}
