% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_gtf}
\alias{.generate_gtf}
\title{Generate Mock GTF File}
\usage{
.generate_gtf(output_file, n_records, options, genome_context = NULL, compress)
}
\description{
Generate Mock GTF File
}
\keyword{internal}
