% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_mtx}
\alias{.generate_mtx}
\title{Generate Mock MTX (Matrix Market) File}
\usage{
.generate_mtx(output_file, n_records, options, compress)
}
\description{
Generate Mock MTX (Matrix Market) File
}
\keyword{internal}
