% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_random_sequence}
\alias{.generate_random_sequence}
\title{Generate Random DNA Sequence}
\usage{
.generate_random_sequence(length)
}
\description{
Generate Random DNA Sequence
}
\keyword{internal}
