% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_realistic_qualities}
\alias{.generate_realistic_qualities}
\title{Generate Realistic Quality Scores}
\usage{
.generate_realistic_qualities(read_length, min_quality, max_quality)
}
\description{
Generate Realistic Quality Scores
}
\keyword{internal}
