% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_sam}
\alias{.generate_sam}
\title{Generate Mock SAM File}
\usage{
.generate_sam(output_file, n_records, options, genome_context = NULL, compress)
}
\description{
Generate Mock SAM File
}
\keyword{internal}
