% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_smart_temp_filename}
\alias{.generate_smart_temp_filename}
\title{Generate Smart Temporary Filename with R1/R2 Support}
\usage{
.generate_smart_temp_filename(datatype, options, compress)
}
\description{
Generate Smart Temporary Filename with R1/R2 Support
}
\keyword{internal}
