% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_temp_filename}
\alias{.generate_temp_filename}
\title{Generate Temporary Filename Based on Datatype}
\usage{
.generate_temp_filename(datatype, compress)
}
\description{
Generate Temporary Filename Based on Datatype
}
\keyword{internal}
