% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.generate_vcf}
\alias{.generate_vcf}
\title{Generate Mock VCF File}
\usage{
.generate_vcf(output_file, n_records, options, genome_context = NULL, compress)
}
\description{
Generate Mock VCF File
}
\keyword{internal}
