% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.get_datatype_options}
\alias{.get_datatype_options}
\title{Get Available Options for Datatype}
\usage{
.get_datatype_options(datatype)
}
\description{
Get Available Options for Datatype
}
\keyword{internal}
