% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.get_shared_context}
\alias{.get_shared_context}
\title{Get Shared Context for Compatible Files (Internal)}
\usage{
.get_shared_context(seed)
}
\description{
Get Shared Context for Compatible Files (Internal)
}
\keyword{internal}
