% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.get_smart_fastq_basename}
\alias{.get_smart_fastq_basename}
\title{Get Smart FASTQ Base Name with R1/R2 Support}
\usage{
.get_smart_fastq_basename(options)
}
\description{
Get Smart FASTQ Base Name with R1/R2 Support
}
\keyword{internal}
