% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{.get_tool_installed_versions}
\alias{.get_tool_installed_versions}
\title{Get Installed Versions (Dynamic)}
\usage{
.get_tool_installed_versions(tool, base_dir)
}
\description{
Get Installed Versions (Dynamic)
}
\keyword{internal}
