% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_tool_version_from_dependencies}
\alias{.get_tool_version_from_dependencies}
\title{Extract the version for the main tool from resolved dependencies}
\usage{
.get_tool_version_from_dependencies(
  tool_name,
  resolved_dependencies,
  fallback_version
)
}
\description{
Extract the version for the main tool from resolved dependencies
}
\keyword{internal}
