% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{.install_tool_direct}
\alias{.install_tool_direct}
\title{Install Tool Directly Using YAML with Proper Version Resolution}
\usage{
.install_tool_direct(toolbox, tool_name, version, config, base_dir)
}
\description{
Install Tool Directly Using YAML with Proper Version Resolution
}
\keyword{internal}
