% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{.install_tool_from_yaml}
\alias{.install_tool_from_yaml}
\title{Install Tool from YAML Environment}
\usage{
.install_tool_from_yaml(toolbox, tool_name, version, config, base_dir)
}
\description{
Install Tool from YAML Environment
}
\keyword{internal}
