% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.introduce_sequencing_errors}
\alias{.introduce_sequencing_errors}
\title{Introduce Sequencing Errors to DNA Sequence}
\usage{
.introduce_sequencing_errors(sequence, error_rate)
}
\description{
Introduce Sequencing Errors to DNA Sequence
}
\keyword{internal}
