% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{.is_installed}
\alias{.is_installed}
\title{Check if Tool Version is Installed (Dynamic)}
\usage{
.is_installed(tool, version = NULL, base_dir = NULL)
}
\description{
Check if Tool Version is Installed (Dynamic)
}
\keyword{internal}
