% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.is_user_modified_toolbox}
\alias{.is_user_modified_toolbox}
\title{Check if Toolbox is User Modified}
\usage{
.is_user_modified_toolbox()
}
\description{
Check if Toolbox is User Modified
}
\keyword{internal}
