% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R, R/utils.R
\name{.normalize_log_level}
\alias{.normalize_log_level}
\title{Normalize Log Level}
\usage{
.normalize_log_level(log_level)

.normalize_log_level(log_level)
}
\arguments{
\item{log_level}{Character or integer. Log level.}
}
\value{
Integer. Numeric log level.
}
\description{
Convert log level to numeric value.
}
\keyword{internal}
