% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{.process_dependencies_for_yaml}
\alias{.process_dependencies_for_yaml}
\title{Process Dependencies for YAML Writing}
\usage{
.process_dependencies_for_yaml(dependencies)
}
\description{
Process Dependencies for YAML Writing
}
\keyword{internal}
