% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.resolve_dependency_versions}
\alias{.resolve_dependency_versions}
\title{Resolve versions for dependencies that don't have explicit versions}
\usage{
.resolve_dependency_versions(dependencies, channels, show_messages = TRUE)
}
\description{
Resolve versions for dependencies that don't have explicit versions
}
\keyword{internal}
