% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{.resolve_latest_version}
\alias{.resolve_latest_version}
\title{Resolve Latest Version for Tool}
\usage{
.resolve_latest_version(config)
}
\description{
Resolve Latest Version for Tool
}
\keyword{internal}
