% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.reverse_complement}
\alias{.reverse_complement}
\title{Generate Reverse Complement of DNA Sequence}
\usage{
.reverse_complement(sequence)
}
\description{
Generate Reverse Complement of DNA Sequence
}
\keyword{internal}
