% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.sn_append_resource_info}
\alias{.sn_append_resource_info}
\title{Append Resource Information to Log}
\usage{
.sn_append_resource_info(log_file, end_time, resources)
}
\arguments{
\item{log_file}{Character. Path to log file.}

\item{end_time}{POSIXct. End time.}

\item{resources}{List. Resource usage information.}
}
\description{
Append Resource Information to Log
}
\keyword{internal}
