% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.sn_create_output_directories}
\alias{.sn_create_output_directories}
\title{Create Output Directories}
\usage{
.sn_create_output_directories(cmd_config, user_params, show_messages = TRUE)
}
\arguments{
\item{cmd_config}{List. Command configuration.}

\item{user_params}{List. User provided parameters.}

\item{show_messages}{Logical. Whether to show creation messages.}
}
\description{
Internal function to create directories for all output files before execution.
}
\keyword{internal}
