% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.sn_determine_log_dir}
\alias{.sn_determine_log_dir}
\title{Determine Log Directory}
\usage{
.sn_determine_log_dir(log_dir, user_params, cmd_config, work_dir)
}
\arguments{
\item{log_dir}{Character. User specified log directory.}

\item{user_params}{List. User parameters.}

\item{cmd_config}{List. Command configuration.}

\item{work_dir}{Character. Working directory.}
}
\value{
Character. Final log directory path.
}
\description{
Determine Log Directory
}
\keyword{internal}
