% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{.sn_discover_installed_tools}
\alias{.sn_discover_installed_tools}
\title{Discover Installed Tools}
\usage{
.sn_discover_installed_tools(toolbox)
}
\arguments{
\item{toolbox}{Toolbox object.}
}
\value{
Updated Toolbox object.
}
\description{
Scan the base directory for installed tools and update toolbox.
}
\keyword{internal}
