% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.sn_display_workflow_info}
\alias{.sn_display_workflow_info}
\title{Display Workflow Information}
\usage{
.sn_display_workflow_info(tool, command, cmd_config, user_params)
}
\arguments{
\item{tool}{Character. Tool name.}

\item{command}{Character. Command name.}

\item{cmd_config}{List. Command configuration.}

\item{user_params}{List. User provided parameters.}
}
\description{
Internal function to display modern workflow-style information about
inputs/outputs/parameters.
}
\keyword{internal}
