% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.sn_extract_parameters}
\alias{.sn_extract_parameters}
\title{Extract Parameters (optional inputs + others)}
\usage{
.sn_extract_parameters(cmd_config, user_params)
}
\arguments{
\item{cmd_config}{List. Command configuration.}

\item{user_params}{List. User parameters.}
}
\value{
List of parameter values.
}
\description{
Extract Parameters (optional inputs + others)
}
\keyword{internal}
