% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.sn_extract_required_inputs}
\alias{.sn_extract_required_inputs}
\title{Extract Required Inputs}
\usage{
.sn_extract_required_inputs(inputs, user_params)
}
\arguments{
\item{inputs}{List. Input definitions.}

\item{user_params}{List. User parameters.}
}
\value{
List of required input parameters.
}
\description{
Extract Required Inputs
}
\keyword{internal}
