% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.sn_extract_user_params}
\alias{.sn_extract_user_params}
\title{Extract User Parameters from Config Section}
\usage{
.sn_extract_user_params(config_section, user_params)
}
\arguments{
\item{config_section}{List. Input or output definitions.}

\item{user_params}{List. User parameters.}
}
\value{
List of parameters that match the config section.
}
\description{
Extract User Parameters from Config Section
}
\keyword{internal}
