% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.sn_find_tool_env}
\alias{.sn_find_tool_env}
\title{Find Tool Environment Path}
\usage{
.sn_find_tool_env(tool, tool_config)
}
\arguments{
\item{tool}{Character. Tool name.}

\item{tool_config}{List. Tool configuration.}
}
\value{
Character. Path to environment or NULL if not found.
}
\description{
Internal function to locate the conda environment for an installed tool.
}
\keyword{internal}
