% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sn_format_file_size}
\alias{.sn_format_file_size}
\title{Format File Size}
\usage{
.sn_format_file_size(bytes)
}
\arguments{
\item{bytes}{Numeric. File size in bytes.}
}
\value{
Character. Formatted file size.
}
\description{
Format File Size
}
\keyword{internal}
