% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sn_format_memory}
\alias{.sn_format_memory}
\title{Format Memory Size}
\usage{
.sn_format_memory(bytes)
}
\arguments{
\item{bytes}{Numeric. Memory size in bytes.}
}
\value{
Character. Formatted memory size.
}
\description{
Format Memory Size
}
\keyword{internal}
