% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{.sn_generate_r_example}
\alias{.sn_generate_r_example}
\title{Generate R Usage Example}
\usage{
.sn_generate_r_example(tool, command, cmd_config)
}
\arguments{
\item{tool}{Character. Tool name.}

\item{command}{Character. Command name.}

\item{cmd_config}{List. Command configuration.}
}
\value{
Character. R code example.
}
\description{
Internal function to generate an R usage example for a command.
}
\keyword{internal}
