% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.sn_normalize_log_level}
\alias{.sn_normalize_log_level}
\title{Normalize Log Level}
\usage{
.sn_normalize_log_level(log_level)
}
\arguments{
\item{log_level}{Character or Integer. Logging level.}
}
\value{
Integer. Normalized log level.
}
\description{
Normalize Log Level
}
\keyword{internal}
