% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-engine.R
\name{.sn_validate_params}
\alias{.sn_validate_params}
\title{Validate Required Parameters}
\usage{
.sn_validate_params(params, inputs)
}
\arguments{
\item{params}{List. Final parameters after merging defaults.}

\item{inputs}{List. Input definitions.}
}
\value{
NULL (throws error if validation fails)
}
\description{
Internal function to check that all required parameters are provided.
}
\keyword{internal}
