% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.validate_and_collect_options}
\alias{.validate_and_collect_options}
\title{Validate and Collect Options from ...}
\usage{
.validate_and_collect_options(datatype, ...)
}
\description{
Validate and Collect Options from ...
}
\keyword{internal}
