% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdata.R
\name{.write_lines}
\alias{.write_lines}
\title{Write Lines with Optional Compression}
\usage{
.write_lines(lines, output_file, compress = FALSE)
}
\description{
Write Lines with Optional Compression
}
\keyword{internal}
