% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null Default Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side value.}

\item{y}{Right-hand side value.}
}
\value{
x if not NULL, otherwise y.
}
\description{
Returns right-hand side if left-hand side is NULL.
}
\keyword{internal}
