% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{print.sn_result}
\alias{print.sn_result}
\title{Print Method for Tool Execution Results}
\usage{
\method{print}{sn_result}(x, ...)
}
\arguments{
\item{x}{An sn_result object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for Tool Execution Results
}
