% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,Toolbox-method}
\alias{show,Toolbox-method}
\title{Show Method for Toolbox}
\usage{
\S4method{show}{Toolbox}(object)
}
\arguments{
\item{object}{A Toolbox object}
}
\description{
Show Method for Toolbox
}
