% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,ToolCall-method}
\alias{show,ToolCall-method}
\title{Show Method for ToolCall}
\usage{
\S4method{show}{ToolCall}(object)
}
\arguments{
\item{object}{A ToolCall object}
}
\description{
Show Method for ToolCall
}
