% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{sn_get_tool}
\alias{sn_get_tool}
\title{Get Tool from Toolbox}
\usage{
sn_get_tool(toolbox, tool_name)
}
\arguments{
\item{toolbox}{Toolbox object}

\item{tool_name}{Character. Tool name}
}
\value{
Tool object
}
\description{
Safely get tool object from toolbox instead of using direct @ access.
}
\examples{
\dontrun{
# Instead of: toolbox@tools[[tool_name]]
tool <- sn_get_tool(toolbox, "samtools")
}
}
\seealso{
Other accessor functions: 
\code{\link{sn_get_installed_versions}()},
\code{\link{sn_get_tool_commands}()},
\code{\link{sn_get_tool_name}()},
\code{\link{sn_get_tool_version}()},
\code{\link{sn_get_tool_versions}()},
\code{\link{sn_get_toolbox_tools}()},
\code{\link{sn_get_toolcall_outputs}()},
\code{\link{sn_get_toolcall_runtime}()},
\code{\link{sn_get_toolcall_status}()},
\code{\link{sn_is_toolcall_success}()}
}
\concept{accessor functions}
