% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{sn_get_tool_commands}
\alias{sn_get_tool_commands}
\title{Get Tool Commands}
\usage{
sn_get_tool_commands(tool)
}
\arguments{
\item{tool}{Tool object}
}
\value{
Character. Available command names
}
\description{
Safely get tool command names instead of using direct @ access.
}
\examples{
\dontrun{
# Instead of: names(tool@commands)
commands <- sn_get_tool_commands(tool)
if ("align" \%in\% commands) {
  # Tool supports align command
}
}
}
\seealso{
Other accessor functions: 
\code{\link{sn_get_installed_versions}()},
\code{\link{sn_get_tool}()},
\code{\link{sn_get_tool_name}()},
\code{\link{sn_get_tool_version}()},
\code{\link{sn_get_tool_versions}()},
\code{\link{sn_get_toolbox_tools}()},
\code{\link{sn_get_toolcall_outputs}()},
\code{\link{sn_get_toolcall_runtime}()},
\code{\link{sn_get_toolcall_status}()},
\code{\link{sn_is_toolcall_success}()}
}
\concept{accessor functions}
