% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{sn_get_tool_name}
\alias{sn_get_tool_name}
\title{Get Tool Name}
\usage{
sn_get_tool_name(tool)
}
\arguments{
\item{tool}{Tool object}
}
\value{
Character. Tool name
}
\description{
Safely get tool name instead of using direct @ access.
}
\examples{
\dontrun{
# Instead of: tool@tool_name
tool_name <- sn_get_tool_name(tool)
}
}
\seealso{
Other accessor functions: 
\code{\link{sn_get_installed_versions}()},
\code{\link{sn_get_tool}()},
\code{\link{sn_get_tool_commands}()},
\code{\link{sn_get_tool_version}()},
\code{\link{sn_get_tool_versions}()},
\code{\link{sn_get_toolbox_tools}()},
\code{\link{sn_get_toolcall_outputs}()},
\code{\link{sn_get_toolcall_runtime}()},
\code{\link{sn_get_toolcall_status}()},
\code{\link{sn_is_toolcall_success}()}
}
\concept{accessor functions}
