% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{sn_get_tool_version}
\alias{sn_get_tool_version}
\title{Get Tool Default Version}
\usage{
sn_get_tool_version(tool)
}
\arguments{
\item{tool}{Tool object}
}
\value{
Character. Default tool version
}
\description{
Safely get tool default version instead of using direct @ access.
}
\examples{
\dontrun{
# Get the default version for a tool
version <- sn_get_tool_version(tool)
}
}
\seealso{
Other accessor functions: 
\code{\link{sn_get_installed_versions}()},
\code{\link{sn_get_tool}()},
\code{\link{sn_get_tool_commands}()},
\code{\link{sn_get_tool_name}()},
\code{\link{sn_get_tool_versions}()},
\code{\link{sn_get_toolbox_tools}()},
\code{\link{sn_get_toolcall_outputs}()},
\code{\link{sn_get_toolcall_runtime}()},
\code{\link{sn_get_toolcall_status}()},
\code{\link{sn_is_toolcall_success}()}
}
\concept{accessor functions}
