% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{sn_get_tool_versions}
\alias{sn_get_tool_versions}
\title{Get All Tool Versions}
\usage{
sn_get_tool_versions(tool)
}
\arguments{
\item{tool}{Tool object}
}
\value{
Character. All available versions
}
\description{
Get all available versions for a tool.
}
\examples{
\dontrun{
# Get all available versions
versions <- sn_get_tool_versions(tool)
}
}
\seealso{
Other accessor functions: 
\code{\link{sn_get_installed_versions}()},
\code{\link{sn_get_tool}()},
\code{\link{sn_get_tool_commands}()},
\code{\link{sn_get_tool_name}()},
\code{\link{sn_get_tool_version}()},
\code{\link{sn_get_toolbox_tools}()},
\code{\link{sn_get_toolcall_outputs}()},
\code{\link{sn_get_toolcall_runtime}()},
\code{\link{sn_get_toolcall_status}()},
\code{\link{sn_is_toolcall_success}()}
}
\concept{accessor functions}
