% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{sn_get_toolbox}
\alias{sn_get_toolbox}
\title{Get Global Toolbox}
\usage{
sn_get_toolbox(refresh = FALSE)
}
\arguments{
\item{refresh}{Logical. Whether to refresh the toolbox.}
}
\value{
Toolbox object.
}
\description{
Get the global cached toolbox instance.
}
