% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{sn_get_toolbox_tools}
\alias{sn_get_toolbox_tools}
\title{Get Toolbox Tools}
\usage{
sn_get_toolbox_tools(toolbox)
}
\arguments{
\item{toolbox}{Toolbox object}
}
\value{
Character. Available tool names
}
\description{
Safely get toolbox tool names instead of using direct @ access.
}
\examples{
\dontrun{
# Instead of: names(toolbox@tools)
tools <- sn_get_toolbox_tools(toolbox)
if ("samtools" \%in\% tools) {
  # Samtools is available in toolbox
}
}
}
\seealso{
Other accessor functions: 
\code{\link{sn_get_installed_versions}()},
\code{\link{sn_get_tool}()},
\code{\link{sn_get_tool_commands}()},
\code{\link{sn_get_tool_name}()},
\code{\link{sn_get_tool_version}()},
\code{\link{sn_get_tool_versions}()},
\code{\link{sn_get_toolcall_outputs}()},
\code{\link{sn_get_toolcall_runtime}()},
\code{\link{sn_get_toolcall_status}()},
\code{\link{sn_is_toolcall_success}()}
}
\concept{accessor functions}
