% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{sn_reset_toolbox}
\alias{sn_reset_toolbox}
\title{Reset Default Toolbox}
\usage{
sn_reset_toolbox()
}
\description{
Reset Default Toolbox
}
\seealso{
Other tool management: 
\code{\link{default_toolbox}},
\code{\link{sn_add_tool}()},
\code{\link{sn_check_user_toolbox}()},
\code{\link{sn_diagnose_tool}()},
\code{\link{sn_get_toolbox_info}()},
\code{\link{sn_initialize_toolbox}()},
\code{\link{sn_install_micromamba}()},
\code{\link{sn_install_tool}()},
\code{\link{sn_load_toolbox}()},
\code{\link{sn_remove_tool}()},
\code{\link{sn_reset_to_default}()},
\code{\link{sn_save_toolbox}()},
\code{\link{sn_validate_tool}()}
}
\concept{tool management}
