% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{sn_load_registry}
\alias{sn_load_registry}
\title{Load Tool Registry}
\usage{
sn_load_registry(tools_dir = NULL)
}
\arguments{
\item{tools_dir}{Character. Directory containing YAML tool definitions.
Defaults to package inst/tools directory.}
}
\value{
List of tool configurations.
}
\description{
Loads the built-in tool registry from YAML files.
}
\seealso{
Other tool information: 
\code{\link{sn_get_tool_config}()},
\code{\link{sn_get_tool_info}()},
\code{\link{sn_help}()},
\code{\link{sn_list_tools}()}
}
\concept{tool information}
